Trait oak_attestation_explain::HumanReadableExplanation

pub trait HumanReadableExplanation {
    // Required method
    fn description(&self) -> Result<String, Error>;
}

Required Methods

fn description(&self) -> Result<String, Error>

Provides human readable explanations for a layer of an attestation data stack, including directions to map the attestation data to source code.